LPS-193641 Refactor AddRelationship modal
