+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectRelationship/ModalAddObjectRelationship.scss
@@ -1,0 +1,12 @@
+.lfr-object-web__modal-add-object-relationship {
+	&-current-object-input {
+		align-items: center;
+		display: flex;
+		gap: 1rem;
+		justify-content: center;
+
+		.form-group {
+			width: 100%;
+		}
+	}
+}
