+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectRelationship/ModalAddObjectRelationship.d.ts
@@ -1,0 +1,20 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+/// <reference types="react" />
+
+interface ModalAddObjectRelationshipProps {
+	handleOnClose: () => void;
+	objectDefinitionExternalReferenceCode: string;
+	objectRelationshipTypes: string[];
+	parameterRequired: boolean;
+}
+export declare function ModalAddObjectRelationship({
+	handleOnClose,
+	objectDefinitionExternalReferenceCode,
+	objectRelationshipTypes,
+	parameterRequired,
+}: ModalAddObjectRelationshipProps): JSX.Element;
+export {};
